feat: Add image upload functionality
This commit adds the necessary code changes to enable image upload functionality to the application. It includes the addition of the "@uploadthing/react" package as a dependency in the package.json file. The "uploadthing.ts" file is also added to the "src/utils" directory, which exports the "UploadButton" and "UploadDropzone" components generated by the "@uploadthing/react" package.

Additionally, the "topnav.tsx" file in the "src/app/_components" directory is modified to include the "UploadButton" component from the "uploadthing" module.

These changes aim to fulfill the task of adding image upload functionality to the application.